Basic Backend Validation

+++ server/routes/bankAccount.js
@@ -8,14 +8,21 @@ const connection = mysql.createConnection({
   
 })
 
-const {hashPassword, isPasswordCorrect, getEmail} = require('../helper-functions/functions')
+const {hashPassword, isPasswordCorrect, getEmail, validateEmail} = require('../helper-functions/functions')
 
 //add bank account
 module.exports.addAccount = async(req,res) => {
 
-    let {account_name, account_type, balance, interest, monthlyFees, user_id} = req.body;
+    let {account_name, account_type, balance, interest, monthlyFees} = req.body;
     let email = getEmail(req.headers.authorization).then((email) => {return email;});
-      //var sql = "INSERT INTO Users (first_name, last_name, email, password, password_salt, phone, profile_image, is_admin) Values ('firstName', 'lastName', 'testingg@testing.com', 'password', 'password_salt', 'phone', 'profile_image', 'is_admin')";
+
+    if(!account_name) return res.status(400).json('Account Name Not Valid');
+    if(!account_type) return res.status(400).json('Account Type Not Valid');
+    if(!balance) return res.status(400).json('Account Balance Not Valid');
+    if(!interest) return res.status(400).json('Account Interest Not Valid');
+    if(!nonthlyFees) return res.status(400).json('Account Fees Not Valid');
+    if(!email) return res.status(400).json('Account Email Not Valid');
+
      var sql = `INSERT INTO Accounts (account_name, account_type, balance, interest, monthlyFees, user_id) Values ('${account_name}', '${account_type}', '${balance}', '${interest}', '${monthlyFees}', (SELECT user_id FROM Users WHERE email = '${await email}'))`;
   
      connection.query(sql, function(err, rows)
@@ -42,6 +49,13 @@ module.exports.editAccount = async(req,res) => {
   
     let {account_name, account_type, balance, interest, monthlyFees} = req.body;
     let email = getEmail(req.headers.authorization).then((email) => {return email;});
+
+    if(!account_name) return res.status(400).json('Account Name Not Valid');
+    if(!account_type) return res.status(400).json('Account Type Not Valid');
+    if(!balance) return res.status(400).json('Account Balance Not Valid');
+    if(!interest) return res.status(400).json('Account Interest Not Valid');
+    if(!nonthlyFees) return res.status(400).json('Account Fees Not Valid');
+    if(!email) return res.status(400).json('Account Email Not Valid');
 
       //var sql = "INSERT INTO Users (first_name, last_name, email, password, password_salt, phone, profile_image, is_admin) Values ('firstName', 'lastName', 'testingg@testing.com', 'password', 'password_salt', 'phone', 'profile_image', 'is_admin')";
      var sql = `Update Accounts SET account_name = '${account_name}', account_type = '${account_type}', balance = '${balance}', interest = '${interest}', monthlyFees = '${monthlyFees}' WHERE account_id = '${account_id}'
@@ -70,6 +84,7 @@ module.exports.deleteAccount = async(req,res) => {
     account_id = req.params.account_id;
 
     let email = getEmail(req.headers.authorization).then((email) => {return email;});
+    if(!email) return res.status(400).json('Account Email Not Valid');
      var sql = `DELETE FROM Accounts WHERE account_id = '${account_id}'
      AND user_id=(SELECT user_id FROM Users WHERE email = '${await email}')`;
      connection.query(sql, function(err, rows)
@@ -90,7 +105,7 @@ module.exports.deleteAccount = async(req,res) => {
 
 module.exports.getAccounts = async(req, res) => {
   let email = getEmail(req.headers.authorization).then((email) => {return email;});
-
+  if(!email) return res.status(400).json('Account Email Not Valid');
   var sql = `SELECT account_id, account_name, account_type, balance, interest, monthlyFees
   FROM Accounts INNER JOIN Users ON Users.user_id = Accounts.user_id
   WHERE email = '${await email}'`;
